refactor: base model 파일 분리
